test(dedupe): consolidate guard_manager tests (5 → 3 files)
Cluster da0ea2376a4d: merge similar_names test files for guard_manager

Files consolidated:
- test_guard_manager.py (NEW): cache + state_collection (131 lines)
- test_guard_manager_execution.py (NEW): cancel_and_safe_run + steps_and_telemetry (211 lines)
- test_guard_manager_runtime.py: renamed from runtime_and_orchestration (147 lines)

Removed:
- test_guard_manager_cache.py
- test_guard_manager_cancel_and_safe_run.py
- test_guard_manager_state_collection.py
- test_guard_manager_steps_and_telemetry.py

All 35 tests pass.
